pyproject.toml
[tool.poetry]
name = "falcon-django"
version = "0.1.0"
description = ""
authors = ["sharofiddin <cyberpancher10@gmail>"]
readme = "README.md"
packages = [{include = "falcon_django"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
django-mptt = "^0.14.0"
pillow = "^9.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
